# Conflicts:
#	backend/src/main/java/net/primepro/primepro/controller/ProductController.java
#	backend/src/main/java/net/primepro/primepro/dto/EmployeeDto.java
#	backend/src/main/java/net/primepro/primepro/dto/ReqRes.java
#	backend/src/main/java/net/primepro/primepro/entity/Employee.java
#	backend/src/main/java/net/primepro/primepro/entity/Inventory.java
#	backend/src/main/java/net/primepro/primepro/entity/OurUsers.java
#	backend/src/main/java/net/primepro/primepro/entity/Product.java
#	backend/src/main/java/net/primepro/primepro/mapper/EmployeeMapper.java
#	backend/src/main/java/net/primepro/primepro/repository/UsersRepo.java
#	backend/src/main/java/net/primepro/primepro/service/EmployeeService.java
#	backend/src/main/java/net/primepro/primepro/service/InventoryService.java
#	backend/src/main/java/net/primepro/primepro/service/ProductService.java
#	backend/src/main/java/net/primepro/primepro/service/UsersManagementService.java
#	backend/src/main/java/net/primepro/primepro/service/impl/EmployeeServiceImpl.java
#	backend/src/main/java/net/primepro/primepro/service/impl/InventoryServiceImpl.java
#	backend/src/main/java/net/primepro/primepro/service/impl/ProductServiceImpl.java

+++ backend/src/main/java/net/primepro/primepro/controller/EmployeeController.java
@@ -42,7 +42,7 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteEmp(@PathVariable("id") Long Id){
+    public String deleteEmp(@PathVariable("id") Integer Id){
         employeeService.deleteEmployee(Id);
         return "Employee deleted";
     }
